custom alerts & more info btn & delete btn + JS

+++ js/timeTrackObj.js
@@ -1,6 +1,13 @@
 function clear() {
-  localStorage.clear();
+  // Usar la función de confirmación personalizada en lugar de window.confirm
+  customConfirm("Do you want to delete all the projects from your Local Storage?", (isConfirmed) => {
+    if (isConfirmed) {
+      localStorage.clear();
+      showCustomAlert("Project Deleted");
+    }
+  });
 }
+
 
 const add = () => {
   recoverProjects();
@@ -12,7 +19,8 @@ const add = () => {
     )
   );
   saveProjects();
-  alert("Project saved");
+  filter();
+  showCustomAlert("Project saved");
 };
 
 const recoverProjects = () => {
@@ -29,9 +37,34 @@ const saveProjects = () => {
   localStorage.setItem("projects", var_json);
 };
 
+const deleteProject = (projectDate) => {
+  recoverProjects();
+  const projectIndex = projects.findIndex((item) => item.date === projectDate);
+  
+  if (projectIndex !== -1) {
+    const projectToDelete = projects[projectIndex];
+    customConfirm(`Do you want to delete ${projectToDelete.name} project from your Local Storage?`, (isConfirmed) => {
+      
+      if (isConfirmed) {
+      projects.splice(projectIndex, 1); 
+      saveProjects();
+      filter(); 
+      showCustomAlert("Project Deleted");
+      }
+    });
+     
+  } else showCustomAlert("Project not found");
+  
+};
+
+
 const chargeSelectsData = () => {
   recoverProjects();
-
+  
+  if (projects.length === 0) {
+    document.getElementById("showProjects").innerHTML = "Storage Empty";
+    return;
+  }
   const filterProjectsByName = [...new Set(projects.map((item) => item.name))];
 
   let listProjectsByName = filterProjectsByName.map(
@@ -58,6 +91,7 @@ const chargeSelectsData = () => {
   ).innerHTML = `<option value="" selected>Select all</option> ${listProjectsByDate.reduce(
     (listado, prod) => listado + prod
   )}`;
+  
 };
 
 const filter = () => {
@@ -92,12 +126,12 @@ const filter = () => {
   ).innerHTML = `<span class="history__span"> Total Time: ${
     calcTotalTime(filteredProjects)
   }</span>`;
-};
 
-const convertToDateTime = (date) => {
-  date = new Date(date);
-  date = date.toLocaleString();
-  return date;
+  document.getElementById(
+    "clear"
+  ).innerHTML = `Delete All Projects`;
+
+
 };
 
 const convertToDate = (date) => {
@@ -124,11 +158,15 @@ const createProjectTable = (filteredProjects) => {
     const date = convertToDate(item.date);
     const time = item.totalTime;
     const projectName = item.name;
+    const svgDel = '<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="history__svg" viewBox="0 0 16 16"><path d="M11 1.5v1h3.5a.5.5 0 0 1 0 1h-.538l-.853 10.66A2 2 0 0 1 11.115 16h-6.23a2 2 0 0 1-1.994-1.84L2.038 3.5H1.5a.5.5 0 0 1 0-1H5v-1A1.5 1.5 0 0 1 6.5 0h3A1.5 1.5 0 0 1 11 1.5Zm-5 0v1h4v-1a.5.5 0 0 0-.5-.5h-3a.5.5 0 0 0-.5.5ZM4.5 5.029l.5 8.5a.5.5 0 1 0 .998-.06l-.5-8.5a.5.5 0 1 0-.998.06Zm6.53-.528a.5.5 0 0 0-.528.47l-.5 8.5a.5.5 0 0 0 .998.058l.5-8.5a.5.5 0 0 0-.47-.528ZM8 4.5a.5.5 0 0 0-.5.5v8.5a.5.5 0 0 0 1 0V5a.5.5 0 0 0-.5-.5Z"/></svg>'
+    const svgInfo = '<svg xmlns="http://www.w3.org/2000/svg" width="16" height="16" fill="currentColor" class="history__svg--info" viewBox="0 0 16 16"><path d="M0 2a2 2 0 0 1 2-2h12a2 2 0 0 1 2 2v12a2 2 0 0 1-2 2H2a2 2 0 0 1-2-2V2zm8.93 4.588-2.29.287-.082.38.45.083c.294.07.352.176.288.469l-.738 3.468c-.194.897.105 1.319.808 1.319.545 0 1.178-.252 1.465-.598l.088-.416c-.2.176-.492.246-.686.246-.275 0-.375-.193-.304-.533L8.93 6.588zM8 5.5a1 1 0 1 0 0-2 1 1 0 0 0 0 2z"/></svg>'
 
     tableHTML += `<tr>
                 <td class="history__td">${date}</td>
                 <td class="history__td">${MiliSecToHourMinSec(time)}</td>
                 <td class="history__td">${projectName}</td>
+                <td class="history__td"><button title="More Info" onclick="moreInfo('${item.date}');">${svgInfo}</button></td>
+                <td class="history__td"><button title="Delete Project" onclick="deleteProject('${item.date}');">${svgDel}</button></td>
                 </tr>`;
   });
 
@@ -137,6 +175,23 @@ const createProjectTable = (filteredProjects) => {
 
   return tableHTML;
 };
+
+const moreInfo = (projectDate) => {
+  recoverProjects();
+  const filterProject = projects.find((item) => item.date === projectDate);
+  
+  if (filterProject) {
+    let finishedTime = new Date(filterProject.date);
+    let startedTime = finishedTime.getTime() - filterProject.totalTime;
+     
+    finishedTime = finishedTime.toLocaleTimeString();
+    startedTime = new Date(startedTime).toLocaleTimeString();
+    showCustomAlert(`Started Time: ${startedTime}\n Finished Time: ${finishedTime}`);
+  } else {
+    showCustomAlert("Project not found");
+  }
+};
+
 
 const calcTotalTime = (filteredProjects) => {
     const totalMilisecs=filteredProjects.reduce(
@@ -157,9 +212,11 @@ const statistics = () => {
 
     //THIS WEEK
     const startOfWeek = new Date();
+
     startOfWeek.setHours(0, 0, 0, 0);
     //getDay()-1 para que la semana empiece en lunes no en domingo
-    startOfWeek.setDate(startOfWeek.getDate() - (startOfWeek.getDay()-1));
+    const dayOfWeek = startOfWeek.getDay() === 0 ? 6 : startOfWeek.getDay() - 1;
+    startOfWeek.setDate(startOfWeek.getDate() - dayOfWeek);
     
     const weekProjects = projects.filter((item) => new Date(item.date) >= startOfWeek && new Date(item.date) <= currentDate);
     document.getElementById("thisWeek").innerHTML = `<div class="statistics__span">THIS WEEK<br>${calcTotalTime(weekProjects)}</div>`;
@@ -168,13 +225,13 @@ const statistics = () => {
      const currentMonth = new Date(currentDate).getMonth();
      // Filtrar proyectos para encontrar los que coinciden con la semana actual
      const monthProjects = projects.filter((item) => new Date(item.date).getMonth() === currentMonth); 
-     document.getElementById("thisMonth").innerHTML = `<div class="statistics__span">THIS MONTH<br>${calcTotalTime(monthProjects)}</div>`;
-
-     
+     document.getElementById("thisMonth").innerHTML = `<div class="statistics__span">THIS MONTH<br>${calcTotalTime(monthProjects)}</div>`; 
 }
 
 
 document.getElementById("filter").onclick = filter;
 document.getElementById("history").onclick = chargeSelectsData;
 document.getElementById("statistics").onclick = statistics;
+document.getElementById("clear").onclick = clear;
+
